+++ src/android/Nabto.java
@@ -66,6 +66,9 @@ public class Nabto extends CordovaPlugin {
         else if (action.equals("setOption")) {
             setOption(args.getString(0), args.getString(1), callbackContext);
         }
+        else if (action.equals("setLocalConnectionPsk")) {
+            setLocalConnectionPsk(args.getString(0), args.getString(1), args.getString(2), callbackContext);
+        }
         else if (action.equals("setBasestationAuthJson")) {
             setBasestationAuthJson(args.getString(0), callbackContext);
         }
@@ -138,6 +141,12 @@ public class Nabto extends CordovaPlugin {
         }
         else if (action.equals("tunnelClose")) {
             tunnelClose(args.getString(0), callbackContext);
+        }
+        else if (action.equals("tunnelSetSendWindowSize")) {
+            tunnelSetSendWindowSize(args.getString(0), args.getString(1), callbackContext);
+        }
+        else if (action.equals("tunnelSetRecvWindowSize")) {
+            tunnelSetRecvWindowSize(args.getString(0), args.getString(1), callbackContext);
         }
         else {
             return false;
@@ -398,6 +407,56 @@ public class Nabto extends CordovaPlugin {
             });
     }
 
+    private byte[] hexstringToBytes(String hexString) throws IllegalArgumentException {
+        int offset;
+        if (hexString.length() == 2*16) {
+            offset = 2;
+        } else if (hexString.length() == 3*16-1) {
+            offset = 3;
+        } else {
+            throw new IllegalArgumentException();
+        }
+        byte[] result = new byte[16];
+        for (int i = 0; i < (hexString.length() + offset-2) / offset; i++) {
+            try {
+                result[i] = (byte)(Integer.parseInt("" + hexString.charAt(offset * i) + hexString.charAt(offset * i + 1), 16));
+            } catch (NumberFormatException e) {
+                throw new IllegalArgumentException(e);
+            }
+        }
+        return result;
+    }
+
+    private void setLocalConnectionPsk(final String host, final String pskId, final String psk, final CallbackContext cc) {
+        final Context context = cordova.getActivity().getApplicationContext();
+        cordova.getThreadPool().execute(new Runnable() {
+                @Override
+                public void run() {
+                    Session initializedSession;
+                    NabtoApi initializedNabto;
+                    synchronized(initMutex) {
+                        if (session == null || nabto == null) {
+                            cc.error(NabtoStatus.API_NOT_INITIALIZED.ordinal());
+                            return;
+                        }
+                        initializedSession = session;
+                        initializedNabto = nabto;
+                    }
+                    NabtoStatus status;
+                    try {
+                        status = initializedNabto.setLocalConnectionPsk(host, hexstringToBytes(pskId), hexstringToBytes(psk), session);
+                    } catch (IllegalArgumentException e) {
+                        status = NabtoStatus.FAILED;
+                    }
+                    if (status != NabtoStatus.OK) {
+                        cc.error(status.ordinal());
+                        return;
+                    }
+                    cc.success();
+                }
+            });
+    }
+    
     private void setBasestationAuthJson(final String authJson, final CallbackContext cc) {
         final Context context = cordova.getActivity().getApplicationContext();
         cordova.getThreadPool().execute(new Runnable() {
@@ -830,4 +889,70 @@ public class Nabto extends CordovaPlugin {
                 }
             });
     }
+
+    private void tunnelSetSendWindowSize(final String tunnelHandle, final String sendWindowSize, final CallbackContext cc) {
+        cordova.getThreadPool().execute(new Runnable() {
+                @Override
+                public void run() {
+                    NabtoApi initializedNabto = getNabto(cc);
+                    if (initializedNabto == null) {
+                        return;
+                    }
+                    int value;
+                    try {
+                        value = Integer.parseInt(sendWindowSize);
+                    } catch (Exception e) {
+                        cc.error(NabtoStatus.FAILED.ordinal());
+                        return;
+                    }
+                    Tunnel tunnel = tunnels.get(tunnelHandle);
+                    if (tunnel != null) {
+                        NabtoStatus status = initializedNabto.tunnelSetSendWindowSize(value, tunnel);
+                        if (status == NabtoStatus.OK) {
+                            tunnels.remove(tunnel);
+                            cc.success();
+                        } else {
+                            cc.error(status.ordinal());
+                        }
+                    } else {
+                        cc.error(NabtoStatus.INVALID_TUNNEL.ordinal());
+                        return;
+                    }
+                }
+            });
+    }
+
+    private void tunnelSetRecvWindowSize(final String tunnelHandle, final String recvWindowSize, final CallbackContext cc) {
+        cordova.getThreadPool().execute(new Runnable() {
+                @Override
+                public void run() {
+                    NabtoApi initializedNabto = getNabto(cc);
+                    if (initializedNabto == null) {
+                        return;
+                    }
+                    int value;
+                    try {
+                        value = Integer.parseInt(recvWindowSize);
+                    } catch (Exception e) {
+                        cc.error(NabtoStatus.FAILED.ordinal());
+                        return;
+                    }
+                    Tunnel tunnel = tunnels.get(tunnelHandle);
+                    if (tunnel != null) {
+                        NabtoStatus status = initializedNabto.tunnelSetRecvWindowSize(value, tunnel);
+                        if (status == NabtoStatus.OK) {
+                            tunnels.remove(tunnel);
+                            cc.success();
+                        } else {
+                            cc.error(status.ordinal());
+                        }
+                    } else {
+                        cc.error(NabtoStatus.INVALID_TUNNEL.ordinal());
+                        return;
+                    }
+                }
+            });
+    }
+
+
 }